ajenti-core/requirements.txt
gevent
gevent-socketio
gevent-websocket
lxml>=2.2.4
passlib
psutil>=0.6.0
python-catcher
python-daemon
requests>=0.12.0
reconfigure>=0.1.43
pyOpenSSL
#Pillow
gipc
pyte
setproctitle
termcolor